Move the brotli and zstd contexts into the shared compression library
The ZlibMode enum, CompressionError, and the BrotliContext and
ZstdContext families (with their encoder/decoder subclasses and JSG
options structs) move from node:zlib into api/compression.{h,c++},
unchanged. They follow Node.js' structure -- node:zlib is their only
consumer today -- but they are compression machinery, not Node
bindings: CompressionStream/DecompressionStream are anticipated to grow
brotli and zstd format support, at which point those formats gain web
frontends over these same contexts.
